requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
pillow==10.4.0
pytesseract==0.3.13
PyMuPDF==1.24.9
python-multipart==0.0.9
requests==2.32.3
rapidfuzz==3.9.7
